requirements.txt
uvicorn[standard]
fastapi[all]~=0.110.1
pydantic~=2.6.4
llama-index
google-generativeai
nebula3-python
PyYAML~=6.0.1
SQLAlchemy~=2.0.29
psycopg2~=2.9.9
